standalone ubuntu oss install script
s3 location

feat: aws-common lib framework

fix: remove signature lib from awsHeaderDate

feat: build canonical request

feat: build signing key hash

feat: AWS signed date and time

feat: eight digit date

fix: import crypto lib

feat: security token lib

feat: write credentials

feat: lambda & configuration
