+++ common/etc/nginx/nginx.conf
@@ -9,15 +9,15 @@ load_module modules/ngx_http_js_module.so;
 load_module modules/ngx_http_xslt_filter_module.so;
 
 # Preserve S3 environment variables for worker threads
-env S3_ACCESS_KEY_ID;
-env S3_SECRET_KEY;
+env AWS_ACCESS_KEY_ID;
+env AWS_SECRET_ACCESS_KEY;
 env S3_BUCKET_NAME;
 env S3_SERVER;
 env S3_SERVER_PORT;
 env S3_SERVER_PROTO;
 env S3_REGION;
 env AWS_SIGS_VERSION;
-env S3_DEBUG;
+env AWS_DEBUG;
 env S3_STYLE;
 env ALLOW_DIRECTORY_LIST;
 env PROVIDE_INDEX_PAGE;
@@ -27,6 +27,9 @@ env PROXY_CACHE_VALID_NOTFOUND;
 env PROXY_CACHE_VALID_FORBIDDEN;
 env HEADER_PREFIXES_TO_STRIP;
 env FOUR_O_FOUR_ON_EMPTY_BUCKET;
+
+env LAMBDA_SERVER;
+env LAMBDA_REGION;
 
 events {
     worker_connections  1024;
