+++ common/etc/nginx/templates/gateway/v4_js_vars.conf.template
@@ -1,4 +1,4 @@
 # This header is needed when doing v4 signature authentication. It
 # specifies the timestamp in which the signature was generated and is used with
 # the x-amz-date header.
-js_set $awsDate s3gateway.awsHeaderDate;
+js_set $awsDate aws_common.awsHeaderDate;
