+++ settings.example
@@ -1,12 +1,12 @@
 S3_BUCKET_NAME=my-bucket
-S3_ACCESS_KEY_ID=ZZZZZZZZZZZZZZZZZZZZ
-S3_SECRET_KEY=aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
+AWS_ACCESS_KEY_ID=ZZZZZZZZZZZZZZZZZZZZ
+AWS_SECRET_ACCESS_KEY=aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
 S3_SERVER=s3-us-east-1.amazonaws.com
 S3_SERVER_PORT=443
 S3_SERVER_PROTO=https
 S3_REGION=us-east-1
 S3_STYLE=virtual
-S3_DEBUG=false
+AWS_DEBUG=false
 AWS_SIGS_VERSION=4
 ALLOW_DIRECTORY_LIST=false
 PROVIDE_INDEX_PAGE=false
