+++ test/docker-compose.yaml
@@ -15,13 +15,13 @@ services:
     restart: "no"
     environment:
       S3_BUCKET_NAME: "bucket-1"
-      S3_ACCESS_KEY_ID: "AKIAIOSFODNN7EXAMPLE"
-      S3_SECRET_KEY: "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
+      AWS_ACCESS_KEY_ID: "AKIAIOSFODNN7EXAMPLE"
+      AWS_SECRET_ACCESS_KEY: "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
       S3_SERVER: "minio"
       S3_SERVER_PORT: "9000"
       S3_SERVER_PROTO: "http"
       S3_REGION: "us-east-1"
-      S3_DEBUG: "true"
+      AWS_DEBUG: "true"
       S3_STYLE: "virtual"
       ALLOW_DIRECTORY_LIST:
       PROVIDE_INDEX_PAGE:
